[WIP] Major refactor: Move routes.py location
